requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka
kafka-python==2.0.2

# MongoDB
pymongo==4.6.0
motor==3.3.2

# ML Libraries
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2
pyod==1.1.2
tensorflow==2.15.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2

# Hugging Face datasets
datasets==2.15.0

